+++ ode/test/test_space.cpp
@@ -90,7 +90,7 @@ static void init_test()
 		      (bounds[i][0] + bounds[i][1])*0.5,
 		      (bounds[i][2] + bounds[i][3])*0.5,
 		      (bounds[i][4] + bounds[i][5])*0.5);
-    dGeomSetData (geom[i],(void*) (i));
+    dGeomSetData (geom[i],(void*)(size_t)(i));
   }
 
   // compute all intersections and put the results in "good_matrix"
